requirements.txt
streamlit
pypdf
google-generativeai
pandas
wordcloud
matplotlib
fpdf
